September 1st, 2014 update
One big change is a mechanism that allows Vampire: the Requiem and
Vampire: the Masquerade to share the same mod, but use different
bookmarks! Want to play Vampire: the Requiem with the rise of the
Invictus in the middle ages, you can! Want to play Dark Ages
Vampire/Vampire the Masquerade in the era of the Baronies of Avalon and
the Tzimisce Voivodate,you can! Want to play the jyhad between the
Camarilla and the Sabbat in the 16th century, you can! Want to play
Roman Camarilla at its height under the Julii, or its fall to the Lancea
et Sanctum? You can! Want to play the War of the Guilds in Stygia in
1096 CE from Wraith: the Oblivion? You can! Pax Hermetica of the Order
of Hermes in 767? You can! Sorcerer's Crusade of 1466? You can! Atleast,
you can when we finish developing all this content. Right now, the two
Requiem bookmarks are playable; and the Masquerade bookmarks will soon
follow. Then we hope to release Requiem for Rome after the Charlemange
DLC and the exciting 480 Migrational Mod (When the World Stopped Making
Sense) are released. Enlil has granted permission for us to use the 480
history/provinces and common/culture files so we can have proper Visgoth
Blood Talon and Vandal Bale Hound werewolves. For a look at our
potential bookmarks, check out our Wiki: bookmarks.